Make gathered resources visible on workers

+++ RTS/Assets/Scripts/Minimap.cs
@@ -72,7 +72,7 @@ public class Minimap : MonoBehaviour
         return worldPosition * ((float)size / mapSize);
     }
 
-    public void SetMinimapElement(Image image, Vector2 positionInWorld, int width, int height, int mapSize)
+    public void SetMinimapElement(Image image, Vector2 positionInWorld, float width, float height, int mapSize)
     {
         image.transform.SetParent(gameObject.transform);
         image.rectTransform.sizeDelta = new Vector2((float)size / mapSize * width, (float)size / mapSize * height);
